rename namespace and log error
