+++ src/SlimData/Program.cs
@@ -1,4 +1,4 @@
-﻿using RaftNode;
+﻿using SlimData;
 
 switch (args.LongLength)
 {
