+++ src/SlimFaas/appsettings.json
@@ -2,7 +2,9 @@
     "Logging": {
         "LogLevel": {
             "Default": "Warning",
-            "Microsoft.AspNetCore": "Warning"
+            "Microsoft.AspNetCore": "Error",
+            "DotNext.Net.Cluster": "Error",
+            "SlimData": "Error"
         }
     },
     "UseKubeConfig": false,
