+++ tests/SlimData.Tests/CommandsTests.cs
@@ -1,5 +1,4 @@
 ﻿using MemoryPack;
-using RaftNode;
 using SlimData.Commands;
 
 namespace SlimData.Tests;
