+++ tests/SlimData.Tests/RaftClusterTests.cs
@@ -14,7 +14,6 @@ using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
-using RaftNode;
 using SlimFaas;
 using SlimFaas.Database;
